Added more flexible templates/module/metainfos settings, closes #13

+++ classes/class.rex_website_manager.inc.php
@@ -5,7 +5,7 @@ class rex_website_manager {
 	
 	public function __construct() {
 		$this->websites = array();
-		$this->currentWebsiteId = 1;
+		$this->currentWebsiteId = rex_website::firstId;
 	}
 
 	public function setCurrentWebsiteId($id) {
@@ -168,6 +168,7 @@ class rex_website_manager {
 		}
 
 		// inludes
+		$initContent .= 'require_once($REX[\'INCLUDE_PATH\'] . \'/addons/website_manager/settings.inc.php\');';
 		$initContent .= 'require_once($REX[\'INCLUDE_PATH\'] . \'/addons/website_manager/classes/class.rex_website.inc.php\');' . PHP_EOL;
 		$initContent .= 'require_once($REX[\'INCLUDE_PATH\'] . \'/addons/website_manager/classes/class.rex_website_manager.inc.php\');' . PHP_EOL . PHP_EOL;
 
@@ -275,6 +276,7 @@ class rex_website_manager {
 	public static function createWebsite($websiteId) {
 		global $REX, $I18N, $curAddonCount, $curPluginCount;
 
+		$sql = rex_sql::factory();
 		$tablePrefix = rex_website::constructTablePrefix($websiteId);
 		$generatedDir = rex_website::constructGeneratedDir($websiteId);
 		$filesDir = rex_website::constructMediaDir($websiteId);
@@ -289,26 +291,84 @@ class rex_website_manager {
 		// database tables
 		// ***************************************************************************************************
 
-		$sql = rex_sql::factory();
+		// users (always identical)
+		$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'user AS SELECT * FROM rex_user');
 
+		// articles (always different)
 		$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'article` ( `pid` int(11) NOT NULL  auto_increment, `id` int(11) NOT NULL  , `re_id` int(11) NOT NULL  , `name` varchar(255) NOT NULL  , `catname` varchar(255) NOT NULL  , `catprior` int(11) NOT NULL  , `attributes` text NOT NULL  , `startpage` tinyint(1) NOT NULL  , `prior` int(11) NOT NULL  , `path` varchar(255) NOT NULL  , `status` tinyint(1) NOT NULL  , `createdate` int(11) NOT NULL  , `updatedate` int(11) NOT NULL  , `template_id` int(11) NOT NULL  , `clang` int(11) NOT NULL  , `createuser` varchar(255) NOT NULL  , `updateuser` varchar(255) NOT NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`pid`)) ENGINE=MyISAM;');
 		$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'article_slice` ( `id` int(11) NOT NULL  auto_increment, `clang` int(11) NOT NULL  , `ctype` int(11) NOT NULL  , `re_article_slice_id` int(11) NOT NULL  ,`value1` text  NULL  , `value2` text NULL  , `value3` text NULL  , `value4` text NULL  , `value5` text NULL  , `value6` text NULL  , `value7` text NULL  , `value8` text NULL  , `value9` text NULL  , `value10` text NULL  , `value11` text NULL  , `value12` text NULL  , `value13` text NULL  , `value14` text NULL  , `value15` text NULL  , `value16` text NULL  , `value17` text NULL  , `value18` text NULL  , `value19` text NULL  , `value20` text NULL  , `file1` varchar(255) NULL  , `file2` varchar(255) NULL  , `file3` varchar(255) NULL  , `file4` varchar(255) NULL  , `file5` varchar(255) NULL  , `file6` varchar(255) NULL  , `file7` varchar(255) NULL  , `file8` varchar(255) NULL  , `file9` varchar(255) NULL  , `file10` varchar(255) NULL  , `filelist1` text NULL  , `filelist2` text NULL  , `filelist3` text NULL  , `filelist4` text NULL  , `filelist5` text NULL  , `filelist6` text NULL  , `filelist7` text NULL  , `filelist8` text NULL  , `filelist9` text NULL  , `filelist10` text NULL  , `link1` varchar(10) NULL  , `link2` varchar(10) NULL  , `link3` varchar(10) NULL  , `link4` varchar(10) NULL  , `link5` varchar(10) NULL  , `link6` varchar(10) NULL  , `link7` varchar(10) NULL  , `link8` varchar(10) NULL  , `link9` varchar(10) NULL  , `link10` varchar(10) NULL  , `linklist1` text NULL  , `linklist2` text NULL  , `linklist3` text NULL  , `linklist4` text NULL  , `linklist5` text NULL  , `linklist6` text NULL  , `linklist7` text NULL  , `linklist8` text NULL  , `linklist9` text NULL  , `linklist10` text NULL  , `php` text NULL  , `html` text NULL  ,`article_id` int(11) NOT NULL  , `modultyp_id` int(11) NOT NULL  , `createdate` int(11) NOT NULL  , `updatedate` int(11) NOT NULL  , `createuser` varchar(255) NOT NULL  , `updateuser` varchar(255) NOT NULL  , `next_article_slice_id` int(11) NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`id`,`re_article_slice_id`,`article_id`,`modultyp_id`)) ENGINE=MyISAM;');
-		$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'clang` ( `id` int(11) NOT NULL  , `name` varchar(255) NOT NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`id`)) ENGINE=MyISAM;');
-		$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'file` ( `file_id` int(11) NOT NULL  auto_increment, `re_file_id` int(11) NOT NULL  , `category_id` int(11) NOT NULL  , `attributes` text NULL  , `filetype` varchar(255) NULL  , `filename` varchar(255) NULL  , `originalname` varchar(255) NULL  , `filesize` varchar(255) NULL  , `width` int(11) NULL  , `height` int(11) NULL  , `title` varchar(255) NULL  , `createdate` int(11) NOT NULL  , `updatedate` int(11) NOT NULL  , `createuser` varchar(255) NOT NULL  , `updateuser` varchar(255) NOT NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`file_id`)) ENGINE=MyISAM;');
-		$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'file_category` ( `id` int(11) NOT NULL  auto_increment, `name` varchar(255) NOT NULL  , `re_id` int(11) NOT NULL  , `path` varchar(255) NOT NULL  , `createdate` int(11) NOT NULL  , `updatedate` int(11) NOT NULL  , `createuser` varchar(255) NOT NULL  , `updateuser` varchar(255) NOT NULL  , `attributes` text NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`id`,`name`)) ENGINE=MyISAM;');
-
-		$sql->setQuery('INSERT INTO `' . $tablePrefix . 'clang` VALUES ("0", "deutsch", 0);');
-
 		$sql->setQuery('ALTER TABLE ' . $tablePrefix . 'article ADD INDEX `id` (`id`), ADD INDEX `clang` (`clang`), ADD UNIQUE INDEX `find_articles` (`id`, `clang`), ADD INDEX `re_id` (`re_id`);');
 		$sql->setQuery('ALTER TABLE ' . $tablePrefix . 'article_slice ADD INDEX `id` (`id`), ADD INDEX `clang` (`clang`), ADD INDEX `re_article_slice_id` (`re_article_slice_id`), ADD INDEX `article_id` (`article_id`), ADD INDEX `find_slices` (`clang`, `article_id`);');
-		$sql->setQuery('ALTER TABLE ' . $tablePrefix . 'file ADD INDEX `re_file_id` (`re_file_id`), ADD INDEX `category_id` (`category_id`);');
-		$sql->setQuery('ALTER TABLE ' . $tablePrefix . 'file_category DROP PRIMARY KEY, ADD PRIMARY KEY (`id`), ADD INDEX `re_id` (`re_id`);');
 
-		$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'user AS SELECT * FROM rex_user');
-		$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'module AS SELECT * FROM rex_module');
-		$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'module_action AS SELECT * FROM rex_module_action');
-		$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'template AS SELECT * FROM rex_template');
-		$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'action AS SELECT * FROM rex_action');
+		// clangs
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_clangs']) {
+			// identical
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'clang AS SELECT * FROM rex_clang');
+		} else {
+			// different
+			$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'clang` ( `id` int(11) NOT NULL  , `name` varchar(255) NOT NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`id`)) ENGINE=MyISAM;');
+			$sql->setQuery('INSERT INTO `' . $tablePrefix . 'clang` VALUES ("0", "deutsch", 0);');
+		}
+
+		// media
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_media']) {
+			// identical
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'file AS SELECT * FROM rex_file');
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'file_category AS SELECT * FROM rex_file_category');
+		} else {
+			// different
+			$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'file` ( `file_id` int(11) NOT NULL  auto_increment, `re_file_id` int(11) NOT NULL  , `category_id` int(11) NOT NULL  , `attributes` text NULL  , `filetype` varchar(255) NULL  , `filename` varchar(255) NULL  , `originalname` varchar(255) NULL  , `filesize` varchar(255) NULL  , `width` int(11) NULL  , `height` int(11) NULL  , `title` varchar(255) NULL  , `createdate` int(11) NOT NULL  , `updatedate` int(11) NOT NULL  , `createuser` varchar(255) NOT NULL  , `updateuser` varchar(255) NOT NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`file_id`)) ENGINE=MyISAM;');
+			$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'file_category` ( `id` int(11) NOT NULL  auto_increment, `name` varchar(255) NOT NULL  , `re_id` int(11) NOT NULL  , `path` varchar(255) NOT NULL  , `createdate` int(11) NOT NULL  , `updatedate` int(11) NOT NULL  , `createuser` varchar(255) NOT NULL  , `updateuser` varchar(255) NOT NULL  , `attributes` text NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`id`,`name`)) ENGINE=MyISAM;');
+			$sql->setQuery('ALTER TABLE ' . $tablePrefix . 'file ADD INDEX `re_file_id` (`re_file_id`), ADD INDEX `category_id` (`category_id`);');
+			$sql->setQuery('ALTER TABLE ' . $tablePrefix . 'file_category DROP PRIMARY KEY, ADD PRIMARY KEY (`id`), ADD INDEX `re_id` (`re_id`);');
+		}
+		
+		// modules
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_modules']) {
+			// identical
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'module AS SELECT * FROM rex_module');
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'module_action AS SELECT * FROM rex_module_action');
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'action AS SELECT * FROM rex_action');
+		} else {
+			// different
+			$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'module` ( `id` int(11) NOT NULL  auto_increment, `name` varchar(255) NOT NULL  , `category_id` int(11) NOT NULL  , `ausgabe` text NOT NULL  , `eingabe` text NOT NULL  , `createuser` varchar(255) NOT NULL  , `updateuser` varchar(255) NOT NULL  , `createdate` int(11) NOT NULL  , `updatedate` int(11) NOT NULL  , `attributes` text NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`id`,`category_id`)) ENGINE=MyISAM;');
+			$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'module_action` ( `id` int(11) NOT NULL  auto_increment, `module_id` int(11) NOT NULL  , `action_id` int(11) NOT NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`id`)) ENGINE=MyISAM;');
+			$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'action` ( `id` int(11) NOT NULL  auto_increment, `name` varchar(255) NOT NULL  , `preview` text NULL  , `presave` text NULL  , `postsave` text NULL  , `previewmode` tinyint(4) NULL  , `presavemode` tinyint(4) NULL  , `postsavemode` tinyint(4) NULL  , `createuser` varchar(255) NOT NULL  , `createdate` int(11) NOT NULL  , `updateuser` varchar(255) NOT NULL  , `updatedate` int(11) NOT NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`id`)) ENGINE=MyISAM;');
+			$sql->setQuery('ALTER TABLE ' . $tablePrefix . 'module DROP PRIMARY KEY, ADD PRIMARY KEY (`id`), ADD INDEX `category_id` (`category_id`);');
+		}
+
+		// templates
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_templates']) {
+			// identical
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . 'template AS SELECT * FROM rex_template');
+		} else {
+			// different
+			$sql->setQuery('CREATE TABLE `' . $tablePrefix . 'template` ( `id` int(11) NOT NULL  auto_increment, `label` varchar(255) NULL  , `name` varchar(255) NULL  , `content` text NULL  , `active` tinyint(1) NULL  , `createuser` varchar(255) NOT NULL  , `updateuser` varchar(255) NOT NULL  , `createdate` int(11) NOT NULL  , `updatedate` int(11) NOT NULL  , `attributes` text NULL  , `revision` int(11) NOT NULL  , PRIMARY KEY (`id`)) ENGINE=MyISAM;');
+		}
+
+		// meta infos
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_meta_infos']) {
+			// identical
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . '62_type AS SELECT * FROM rex_62_type');
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . '62_params AS SELECT * FROM rex_62_params');
+		} else {
+			// different (addon will be reinstalled)
+			if (!in_array('metainfo', $REX['WEBSITE_MANAGER_SETTINGS']['reinstall_addons'])) {
+				$REX['WEBSITE_MANAGER_SETTINGS']['reinstall_addons'][] = 'metainfo';
+			}
+		}
+
+		// image types
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_image_types']) {
+			// identical
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . '679_types AS SELECT * FROM rex_679_types');
+			$sql->setQuery('CREATE VIEW ' . $tablePrefix . '679_type_effects AS SELECT * FROM rex_679_type_effects');
+		} else {
+			// different (addon will be reinstalled)
+			if (!in_array('image_manager', $REX['WEBSITE_MANAGER_SETTINGS']['reinstall_addons'])) {
+				$REX['WEBSITE_MANAGER_SETTINGS']['reinstall_addons'][] = 'image_manager';
+			}
+		}
 
 		// ***************************************************************************************************
 		// directories
@@ -322,14 +382,16 @@ class rex_website_manager {
 		mkdir($includePath . $generatedDir . '/files', $REX['DIRPERM']);
 		mkdir($includePath . $generatedDir . '/templates', $REX['DIRPERM']);
 
-		mkdir($htdocsPath . $filesDir, $REX['DIRPERM']);
+		if (!$REX['WEBSITE_MANAGER_SETTINGS']['identical_media']) {
+			mkdir($htdocsPath . $filesDir, $REX['DIRPERM']);
+		}
 
 		// ***************************************************************************************************
 		// addons/plugins
 		// ***************************************************************************************************
 
-		$reinstallAddons = $REX['ADDON']['website_manager']['settings']['reinstall_addons'];
-		$reinstallPlugins = $REX['ADDON']['website_manager']['settings']['reinstall_plugins'];
+		$reinstallAddons = $REX['WEBSITE_MANAGER_SETTINGS']['reinstall_addons'];
+		$reinstallPlugins = $REX['WEBSITE_MANAGER_SETTINGS']['reinstall_plugins'];
 
 		// set stuff for new website
 		$REX['TABLE_PREFIX'] = $tablePrefix;
@@ -370,6 +432,7 @@ class rex_website_manager {
 	public static function destroyWebsite($websiteId) {
 		global $REX, $I18N;
 
+		$sql = rex_sql::factory();
 		$tablePrefix = rex_website::constructTablePrefix($websiteId);
 		$generatedDir = rex_website::constructGeneratedDir($websiteId);
 		$filesDir = rex_website::constructMediaDir($websiteId);
@@ -381,22 +444,52 @@ class rex_website_manager {
 		}
 
 		// ***************************************************************************************************
+		// database views
+		// ***************************************************************************************************
+
+		// users
+		$sql->setQuery('DROP VIEW ' . $tablePrefix . 'user');
+
+		// clangs
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_clangs']) {
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . 'clang');
+		} 
+
+		// media
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_media']) {
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . 'file');
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . 'file_category');
+		}
+		
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_modules']) {
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . 'module');
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . 'module_action');
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . 'action');
+		}
+
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_templates']) {
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . 'template');
+		}
+
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_meta_infos']) {
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . '62_type');
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . '62_params');
+		}
+
+		if ($REX['WEBSITE_MANAGER_SETTINGS']['identical_image_types']) {
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . '679_types');
+			$sql->setQuery('DROP VIEW ' . $tablePrefix . '679_type_effects');
+		}
+
+		// ***************************************************************************************************
 		// database tables
 		// ***************************************************************************************************
 
-		$sql = rex_sql::factory();
+		$tables = $sql->showTables(1, $tablePrefix);
 
-		$sql->setQuery('DROP TABLE ' . $tablePrefix . 'article');
-		$sql->setQuery('DROP TABLE ' . $tablePrefix . 'article_slice');
-		$sql->setQuery('DROP TABLE ' . $tablePrefix . 'clang');
-		$sql->setQuery('DROP TABLE ' . $tablePrefix . 'file');
-		$sql->setQuery('DROP TABLE ' . $tablePrefix . 'file_category');
-
-		$sql->setQuery('DROP VIEW ' . $tablePrefix . 'user');
-		$sql->setQuery('DROP VIEW ' . $tablePrefix . 'module');
-		$sql->setQuery('DROP VIEW ' . $tablePrefix . 'module_action');
-		$sql->setQuery('DROP VIEW ' . $tablePrefix . 'template');
-		$sql->setQuery('DROP VIEW ' . $tablePrefix . 'action');
+		for ($i = 0; $i < count($tables); $i++) {
+			$sql->setQuery('DROP TABLE ' . $tables[$i]);
+		}
 
 		// ***************************************************************************************************
 		// directories
@@ -406,16 +499,9 @@ class rex_website_manager {
 		$htdocsPath = realpath($REX['HTDOCS_PATH']) . DIRECTORY_SEPARATOR;
 
 		rex_website_manager_utils::rrmdir($includePath . $generatedDir);
-		rex_website_manager_utils::rrmdir($htdocsPath . $filesDir);
 
-		// ***************************************************************************************************
-		// addons
-		// ***************************************************************************************************
-
-		$tables = $sql->showTables(1, $tablePrefix);
-
-		for ($i = 0; $i < count($tables); $i++) {
-			$sql->setQuery('DROP TABLE ' . $tables[$i]);
+		if (!$REX['WEBSITE_MANAGER_SETTINGS']['identical_media']) {
+			rex_website_manager_utils::rrmdir($htdocsPath . $filesDir);
 		}
 	}
 }
